socketpair use AF_UNIX if no-windows os

+++ libhttp/source/http-server-route.cpp
@@ -39,7 +39,7 @@ int http_server_delroute(const char* path)
 int http_server_route(void* http, http_session_t* session, const char* method, const char* path)
 {
 	char reqpath[1024];//[PATH_MAX];
-	struct uri_t* uri = uri_parse(path, strlen(path));
+	struct uri_t* uri = uri_parse(path, (int)strlen(path));
 	url_decode(uri->path, -1, reqpath, sizeof(reqpath));
 	uri_free(uri);
